feat: add Terraform job execution with Dagger integration
Add comprehensive support for executing Terraform commands within Dagger
containers through new job abstractions. Introduce JobTerraform and
JobTerraformExec methods on the Infra type to configure and run Terraform
operations in a flexible, reusable manner.

JobTerraform sets up a Dagger container preconfigured with Terraform modules,
AWS credentials, environment variables, SSH authentication, and optional tools
like TFLint and terraform-docs. It supports loading .env files, disabling cache,
and injecting tokens for GitHub, GitLab, and Terraform Registry. This method
returns a ready-to-use container for further command execution.

JobTerraformExec builds on JobTerraform by additionally executing a specified
Terraform command with arguments inside the prepared container. It returns the
command output or an error. This separation allows reuse of container setup
logic and simplifies running arbitrary Terraform commands.

The implementation improves modularity and maintainability by centralizing
Terraform container configuration and command execution. It leverages existing
utility functions for environment variable handling, path resolution, and secret
management. The approach ensures consistent environment setup across Terraform
jobs and reduces duplication.

Affected files include pipeline/infra/terraform_jobs.go and pipeline/infra/utils.go,
where helper functions and types were added or extended to support these jobs.

This change preserves backwards compatibility by adding new methods without
modifying existing interfaces. It fixes prior
